+++ apps/src/componentLibrary/dropdown/checkboxDropdown/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.3.0](https://github.com/code-dot-org/code-dot-org/pull/57827)
 * added `thick` and `thin` label style types (`labelType` prop)
+* added black and white background color fill
 
 ## [0.2.0](https://github.com/code-dot-org/code-dot-org/pull/56683)
 * moved `CheckboxDropdown` to -> `dropdown/checkboxDropdown`